Rust-ci-statistics/requirements.txt
gitpython
tabulate
dotenv
pandas
seaborn
matplotlib
torch
tensorflow
flax
transformers
frontmatter